[hotfix][table-api] Relocate and document Over window offsets
