+++ flink-table/flink-table-planner/src/main/scala/org/apache/flink/table/expressions/overOffsets.scala
@@ -1,0 +1,53 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.table.expressions
+
+import org.apache.flink.table.typeutils.{RowIntervalTypeInfo, TimeIntervalTypeInfo}
+
+case class CurrentRow() extends PlannerExpression {
+  override private[flink] def resultType = RowIntervalTypeInfo.INTERVAL_ROWS
+
+  override private[flink] def children = Seq()
+
+  override def toString = "CURRENT ROW"
+}
+
+case class CurrentRange() extends PlannerExpression {
+  override private[flink] def resultType = TimeIntervalTypeInfo.INTERVAL_MILLIS
+
+  override private[flink] def children = Seq()
+
+  override def toString = "CURRENT RANGE"
+}
+
+case class UnboundedRow() extends PlannerExpression {
+  override private[flink] def resultType = RowIntervalTypeInfo.INTERVAL_ROWS
+
+  override private[flink] def children = Seq()
+
+  override def toString = "UNBOUNDED ROW"
+}
+
+case class UnboundedRange() extends PlannerExpression {
+  override private[flink] def resultType = TimeIntervalTypeInfo.INTERVAL_MILLIS
+
+  override private[flink] def children = Seq()
+
+  override def toString = "UNBOUNDED RANGE"
+}
